解决了快速自动关机不能区分日期的bug

+++ Main.py
@@ -15,15 +15,15 @@ class Infomation:
                     tmp = each.replace(" ", "").replace("\n", "")
                     tmp = tmp.split("=")
                     setattr(self, tmp[0], int(tmp[1]))
-            if self.restTime <= 0:
+            #对当前时间的记录，用于校准
+            nowTime = time.localtime(time.time())
+            self.nowMinutes = nowTime.tm_hour * 60 + nowTime.tm_min
+            if self.restTime <= 0 and nowTime.tm_mday == self.day:
                 import os
                 if os.name.find('nt') >= 0:
                     os.system("shutdown -s -t 0")
                 else:
                     os.system("shutdown now")
-            #对当前时间的记录，用于校准
-            nowTime = time.localtime(time.time())
-            self.nowMinutes = nowTime.tm_hour * 60 + nowTime.tm_min
         else:
             nowDate = time.localtime(time.time())
             self.year = nowDate.tm_year
